+++ tests/test_codex_smart_routing_v2.py
@@ -10,31 +10,6 @@ from ucode.smart_routing import codex_interposer, codex_routing, v2
 WS = "https://example.databricks.com"
 
 
-class TestCodexConfigArgs:
-    def test_layers_provider_overrides_without_replacing_user_config(self, monkeypatch):
-        monkeypatch.setattr(codex, "ucode_version", lambda: "0.1.0")
-        monkeypatch.setattr(codex, "agent_version", lambda binary: "0.148.0")
-
-        overlay = codex.render_overlay(
-            WS,
-            "gpt-5.6-luna",
-            "myprof",
-        )
-        args = v2.codex_config_args(overlay)
-
-        assert args[:4] == [
-            "--config",
-            'model_provider="ucode-databricks"',
-            "--config",
-            'model="gpt-5.6-luna"',
-        ]
-        provider_override = args[-1]
-        assert provider_override.startswith("model_providers.ucode-databricks={")
-        assert "/ai-gateway/codex/v1" in provider_override
-        assert 'command = "' in provider_override
-        assert '"myprof"' in provider_override
-
-
 def test_smart_routing_switch_message_is_boxed():
     message = v2.format_routing_notice("model-x", "Because X.")
